revert(intercom): restore tool-driven connections
Stop foreground subagent launches and bridged child startup from importing or connecting Intercom automatically. Keep lazy lifecycle replay and foreground detach handling intact, and document that models should connect through the Intercom tool when coordination is needed.

Assistant-model: GPT-5.6 Sol
